Refactor BitFlags to use IBitFlags interface
Refactored EntityId to accept IBitFlags for compIdsMask.

Updated IsAlive property to check if compIdsMask is not null and not empty.

Changed Filter class to use BitFlags128 for incIds and excIds.

Replaced BitFlags64 with IBitFlags interface.

Updated BitFlags128 to implement IBitFlags interface.

Adjusted BitFlags128 equality operators to compare with IBitFlags.

Modified HasAnyFlag and HasAllFlags methods in BitFlags128 to accept IBitFlags parameter.